Fix #141
Fixed typos in the marker of the end of the K8S block for removal.

Fixed pipeline tests to not assume existence of either K8S or CF
scripts.

Fixed JobScripts tests to only verify availability of the CF/K8S jobs in
case of them not being excluded during the customization.